+++ ui/mods/keybind_extensions/keybinds.js
@@ -76,6 +76,7 @@
   set = 'gameplay'
   display_sub_group = 'Selection Edit'
   kb('only_one_in_selection')
+  kb('halve_selection')
   kb('only_construction_in_selection')
   kb('remove_construction_from_selection')
   kb('only_bots_in_selection')